# apps/ebpco-api/contract/citizen-endpoints.openapi.yaml
...
openapi: 3.1.0
info:
  title: eBPCO citizen endpoints — permit, documents, resubmission
  version: 1.0.0
  description: |
    Shipped in eBPCOBackend: C-1 at f7eb40e, C-2 at 5a0f18a, D-8 at 18028a8.

    All three require a citizen bearer token. All three answer 404 for an
    application that is not the caller's, which is the SAME 404 as one that does
    not exist: a reference number that resolves differently would confirm that a
    neighbour has applied.
paths:
...
  /applications/{applicationId}/permit:
    get:
      operationId: getPermit
      summary: The issued permit — number, date, scope, conditions, collection
      security: [{ bearerAuth: [applications:read] }]
      parameters:
        - { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: The permit.
          content: { application/json: { schema: { $ref: '#/components/schemas/Permit' } } }
        '404':
          description: |
            Not the caller's application, no such application, OR the permit has
            not been issued yet. The first two are deliberately indistinguishable;
            the third says so in `detail` ("No permit has been issued for this
            application yet"), because an applicant reading their OWN application
            is entitled to know it simply is not ready.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  /applications/{applicationId}/documents:
    get:
      operationId: listApplicationDocuments
      summary: Every document on the application, and the office's verdict on each
      description: |
        A BARE ARRAY, not an envelope — the same shape as the timeline. Empty
        array for an application of theirs with nothing uploaded, which is a
        different answer from 404 and a true one.

        This is the only route that returns a document id, which content
        retrieval and resubmission both require.
      security: [{ bearerAuth: [applications:read] }]
      parameters:
        - { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: The documents, oldest upload first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApplicationDocument' }
        '404':
          description: Not the caller's application, or no such application.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  /applications/{applicationId}/requirements:
    get:
      operationId: getApplicationRequirements
      summary: The checklist this application was judged against, and what answers it
      description: |
        Read from the snapshot taken at filing, NOT the live catalogue: the
        checklist changes and a filed application must not.

        `unattributedDocuments` travels WITH the list and must be rendered with
        it. A document with no `requirementCode` may well answer a requirement —
        nobody recorded which — so a `not-provided` entry is only certainly
        missing when `attributionComplete` is true. Showing "3 missing" while
        hiding "and 7 documents nobody attributed" tells an applicant to send
        documents they already sent.
      security: [{ bearerAuth: [applications:read] }]
      parameters:
        - { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: The checklist, with what answers each entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationRequirements' }
        '404':
          description: Not the caller's application, or no such application.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Permit:
      type: object
      additionalProperties: false
      required: [permitNumber, issuedDate, scope, conditions, release]
      properties:
        permitNumber: { type: string, examples: ['FP-2026-000001'] }
        issuedDate: { type: string, format: date-time }
        scope:
          type: [string, 'null']
          description: What the permit covers. Null when the officer recorded none.
        conditions:
          type: array
          items: { type: string }
          description: |
            What the permit REQUIRES of the holder — a cash bond, a setback,
            notice before excavation. Render in full: a citizen who cannot read
            these cannot comply with them. Empty array when there are none.
        release:
          description: |
            ALWAYS PRESENT, and null until an officer has prepared the release.
            Null is a fact to render ("not yet ready to collect"), not an absent
            key — both branches are recorded:
            `applicant.applications.permit` (populated) and
            `applicant.applications.permit.beforeRelease` (null).
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [status, method, releasedAt]
              properties:
                status: { type: string, enum: ['Not Ready', 'Ready for Release', 'Released'] }
                method: { type: [string, 'null'], enum: ['Physical Claim', 'Authorized Representative', null] }
                releasedAt: { type: [string, 'null'], format: date-time }
    ApplicationDocument:
      type: object
      additionalProperties: false
      required:
        [id, label, fileName, contentType, byteSize, sha256, uploadedAt, expiresOn,
         certifiedOn, requirementCode, reviewStatus, reviewedAt, reviewReason,
         reviewRemark, supersedesDocumentId, supersededByDocumentId, scanCleared,
         quarantined]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        fileName: { type: string }
        contentType: { type: string }
        byteSize:
          type: string
          description: A STRING, not a number — the column is a bigint and JSON numbers lose precision.
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        uploadedAt: { type: string, format: date-time }
        expiresOn: { type: [string, 'null'], format: date }
        certifiedOn:
          type: [string, 'null']
          format: date
          description: |
            When the ISSUING OFFICE certified this document, read off its face.

            Required by the Municipal ruling of 2026-09-03: the admin note on a
            reused document says it is reused **and the date it was certified**.
            An expiry cannot answer that, and an upload date is when the file
            reached this service — three different dates, and none may stand in
            for another. A test asserts they are distinct.

            **NULL MEANS NOT RECORDED**, never that the document was
            uncertified. Nobody supplies it yet — who does is a workflow
            decision with the Municipality — so it is null on every document
            today. Render "certification date not recorded"; do NOT fall back to
            the upload date, which would tell an officer a document was
            certified on the day it reached us.
          examples: ['2024-03-12']
        requirementCode:
          type: [string, 'null']
          description: |
            Which checklist entry this document answers. NULL MEANS NOT
            ATTRIBUTED — nobody recorded which requirement it is for — and never
            that it answers none. Every document uploaded before this existed
            carries null.
        reviewStatus:
          type: [string, 'null']
          enum: ['Missing', 'Uploaded', 'Submitted', 'Under Review', 'Accepted',
                 'Rejected', 'Revision Required', 'Expired', null]
          description: |
            The OFFICER's verdict. NULL MEANS NOBODY HAS LOOKED YET — it does not
            mean nothing is wrong. Rendering null as a tick would tell an
            applicant their document passed when it has not been opened.
        reviewedAt: { type: [string, 'null'], format: date-time }
        reviewReason:
          description: |
            The standard, reusable reason. Null when none was cited. USE `label`
            for display rather than keeping your own copy of the catalogue: the
            LGU can edit it, and a client-side copy drifts. `code` is what to
            switch on.
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [code, label, description]
              properties:
                code: { type: string, examples: ['illegible'] }
                label: { type: string, examples: ['Illegible'] }
                description: { type: string }
        reviewRemark:
          type: [string, 'null']
          description: Free text written for THIS applicant about THIS document.
        supersedesDocumentId: { type: [string, 'null'], format: uuid }
        supersededByDocumentId: { type: [string, 'null'], format: uuid }
        scanCleared:
          type: boolean
          description: Malware scan completed and passed. A DIFFERENT axis from reviewStatus.
        quarantined:
          type: boolean
          description: |
            Held by the malware scanner. Not an evaluation outcome — an officer's
            rejection does not mean a virus, and a quarantined file is not a
            verdict on the application.
    ApplicationRequirements:
      type: object
      additionalProperties: false
      required: [requirements, unattributedDocuments, attributionComplete]
      properties:
        requirements:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [code, label, description, required, documentIds, status]
            properties:
              code: { type: string }
              label: { type: string }
              description: { type: string }
              required:
                type: boolean
                description: False for a document asked for only in some circumstances. It still appears.
              documentIds:
                type: array
                items: { type: string, format: uuid }
                description: |
                  The CURRENT documents answering this entry. A superseded
                  document is excluded, so a replaced-and-accepted requirement
                  does not still read as rejected because of what it replaced.
              status:
                type: string
                enum: [provided, not-provided]
                description: |
                  'provided' means a document is ATTRIBUTED to this entry. It
                  does NOT mean an officer accepted it — that is `reviewStatus`
                  on the document. Conflating them would tell an applicant their
                  rejected lot plan satisfies the requirement it failed.
        unattributedDocuments:
          type: integer
          description: Documents on this application carrying no requirementCode.
        attributionComplete:
          type: boolean
          description: |
            THE FIELD THAT SAYS WHETHER 'not-provided' CAN BE TRUSTED. False
            means some documents carry no code, so an entry reported
            'not-provided' may in fact have been answered by one of them.
...
    Problem:
      type: object
      description: RFC 9457. `detail` is the human-readable half.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }